+++ src/main/java/net/snowflake/ingest/streaming/internal/serialization/DuplicateKeyValidatedObject.java
@@ -2,10 +2,10 @@
  * Copyright (c) 2025 Snowflake Computing Inc. All rights reserved.
  */
 
-package net.snowflake.ingest.streaming.internal;
+package net.snowflake.ingest.streaming.internal.serialization;
 
 /**
- * A wrapper for an Object that is going to be sanitized by {@link
+ * A wrapper for an Object that is going to be validated by {@link
  * DuplicateKeyValidatingSerializer}.
  */
 public class DuplicateKeyValidatedObject {
